Restricts prefpane unit tests to local architecture.
Without this fix, building with Xcode 3.1 on a Mac Pro running 10.9.5
was attempting to include a ppc build and failing.  There was already
a config file that purported to fix this just to avoid running the
tests under Rosetta (back when Rosetta existed), but it wasn't working.

This fix precludes running both 32- and 64-bit builds of the tests (and
actually winds up running in 32-bit mode, even on a 64-bit machine), but
Xcode seems to lack sufficient flexibility to get this right.